feat(Feedback): refactor feedback components and fix import paths
